+++ functions.php
@@ -160,7 +160,7 @@ add_action( 'after_setup_theme', function() {
         'core/textColumns',
       ],
       'page' => [
-        'core/media-text',
+        'all',
       ],
     ],